Log ETag of extracted cache archive if available
This commit adds the ETag from the downloaded cache file to the
"Downloading cache" message in the job.

When debugging CI jobs, we found that it would be useful to log the
ETag of the extracted cache file.The ETag header is supported by
Amazon S3, Google Cloud Storage, and Azure.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/460139

Changelog: added