Allow callers of Kubernetes log processor to wait for Goroutines
The upgrade to Go 1.22 detected a race condition between the
`logProcessor.scan()` and the mocked `logStreamer.Stream()` method.
The race condition flagged an issue with the third argument of the
`Stream()` method, which is a `io.PipeWriter`. The race condition
happened because:

1. `scan()` can still be stuck reading from the pipe when the context
    is canceled.
2. When the pipe is closed, the call in `scan()` to
   `logsScanner.reader.ReadString()` will detect a broken pipe and
   attempt to call `readCloseError()` to write the `onceError` on
   the `io.PipeWriter`.
3. The `AssertExpectations()` method will attempt to inspect the
   `io.PipeWriter` and cause a data race with step 2.

To avoid this race, introduce `logProcessor.Finalize()` that will
return once all Goroutines have completed.

Changelog: fixed