Fix Windows killer graceful termination and related tests
When using the new (non-legacy) process strategy CreateProcess is invoked
with CREATE_NEW_PROCESS_GROUP which implicitly results in Ctrl+C
being disabled for the process.

Generate Ctrl+Break instead which has similar purpose and can be targeted
at a process group (which Ctrl+C can not).

Since Powershell (when run interactively) handles Ctrl+Break to start
a debugger make sure the secondary sub-shell is started with
-NonInteractive (like the primary shell already is).

This does suppress the debugger for "pwsh" but not the old "powershell".

Revise the sleep binary used by the integration test to make it easier
to distinguish how the killed process was terminated.

Previously the kill tests passed even when the sleep process
died of natural causes.