Make pull manager allow for different images
In order to allow for different pull policies for e.g. the services
images, we need to have the pull manager allow different support that.
We therefor need to have the pull manager track different policies for
different images, and we need to provide those different policies when
creating the manager.

The helper image does not have a separate pull policy from the build
image, thus we copy the pull manager config for the build image and use
it also for the helper image.