Use correct build image in TestPrepare tests
Different test cases define the image name either on the Build or in the
Runner config, but we only use the one from the build to determine the
pull policy. Previously, that worked, because the pull manager only kept
track of one image and it's pull policies. Now, that it tracks multiple
images, we need to be certain, which image we care about.

This is a hotfix, to make the tests succeed. Generally we should see to
not put such logic into tests.