Have the k8s executor successfully discover the windows version
When `HelperImageAutosetArchAndOS` is set, the k8s executor needs to
know which version of windows it runs on, to be able to select the
correct helper image.

Note: This is only necessary when the runner runs on Windows, for Linux
we don't have to know the exact kernel version to be able to select the
correct helper image.