Forward entrypoint logs to the build log
This only happens, when the FF_KUBERNETES_HONOR_ENTRYPOINT is set.

We tail the logs from the k8s API and forward them into the build
logger. When we see a step starting, we pause the forwarding, to avoid
duplicate log entries. When we see a step terminating, we continue to
forward logs again.