Prepare an image with an interesting entrypoint
We will use this image to test certain aspects of honoring the
entrypoint of an image when used with the kubernetes executor.

It generates some logs, we can assert on, before it runs the CMD, and
logs some more things after the CMD has ran, with the help of a trap.