Fix official docker client unit tests
CVE-2024-41110 and GHSA-v23v-6jw2-98fq required upgrading the versions
of github.com/docker/docker and github.com/docker/cli to v25.0.6. The
docker client implementation changed enough in the jump from v24 to v25
that it broke these tests.

Specifically, github.com/docker/docker/client.New overwrites the
http.Client's transport (originally http.Transport) with an
otelhttp.Transport. The later type does not expose the TLSClientConfig
field, which we need to access to assert in the test below that TLS was
set correctly in various scenarios.

To fix this we create the http.Transport instance in newOfficialDockerCLient and
cache it the officialDockerClient instance.


Relevant code from NewClientWithOpts:

    if tr, ok := c.client.Transport.(*http.Transport); ok {
            // Store the base transport before we wrap it in tracing libs below
            // This is used, as an example, to close idle connections when the client is closed
            c.baseTransport = tr  // <-- the original http.Client's Transport
    }

    if c.scheme == "" {
            // TODO(stevvooe): This isn't really the right way to write clients in Go.
            // `NewClient` should probably only take an `*http.Client` and work from there.
            // Unfortunately, the model of having a host-ish/url-thingy as the connection
            // string has us confusing protocol and transport layers. We continue doing
            // this to avoid breaking existing clients but this should be addressed.
            if c.tlsConfig() != nil {
                    c.scheme = "https"
            } else {
                    c.scheme = "http"
            }
    }

    c.client.Transport = otelhttp.NewTransport( // <-- the new one
            c.client.Transport,
            otelhttp.WithTracerProvider(c.tp),
            otelhttp.WithSpanNameFormatter(func(_ string, req *http.Request) string {
                    return req.Method + " " + req.URL.Path
            }),
    )