Fix setting a container's MacAddress
The bump in the docker SDK TO v25 , a container's mac-address was configured via
container.Config.MacAddress. In SDK v25 that approach no longer works.

In the docker SDK v24, a container's mac-address was configured via
container.Config.MacAddress. In SDK v25 that approach has two problems:
- Various linters complain that container.Config.MacAddress is obsolete
  and should not be used.
- If API version >= 1.44 is used, container.Config.MacAddress will have
  no effect at all.

With SDK v25, the right way to set a container's mac-address is via
network.NetworkConfig.EndpointConfig for each network. However, this
approach also requires API >= 1.44.

Requiring API >= 1.44 is not acceptable, and we have to preserve
backward compatibility with API < 1.44. To do this we have to check the
server's API version and:
- if it's < 1.44, configure mac-address the way we used to.
- if it's >= 1.44 configure mac-address via NerworkConfig.

The new implementation of executor.networkConfig attempts to minimize
the changes required to the integration tests to make it pass again.

There are two test cases that changed significantly ("host, *"); where
this configuration used to result in this error when creating the container:

    conflicting options: mac-address and the network mode

Now it succeeds, and the container's NetworkSettings' `host` entry
contains the configured mac-address.

This is, IMO, an expansion of capabilities so it does not represent a
breaking change.

Note the removal of the assertion

    assert.Equal(t, macAddress, info.Config.MacAddress, "config")

from all test cases. This member is also deprecated and while it is
currently still being set, it will be removed in the future, so we may
as well stop relying on it now.