Don't use creds in any git URL
For auth with git, we previously used to add the basic auth creds to the
various URLs (clone URL, insteadOf URLs, ...).

Now, we switch to setting up the basic auth header once, and set it
via `git config` before we do the first pull.

The resulting (base64 encoded) creds will be masked if it ever where to
show up in the build's log.

A side-effect of this is, that in some places where we built up and used
git configs like
`url.https://<user>:<pass>@some.tld.insteadOf=https://some.tld` we can
now drop those; we only need those `insteafOf` stanzas if the URLs are
actually different (e.g. we want to force https instead of SSH).