Introduce DualStack and Accelerate S3 options
The previous Minio S3 client enabled dual-stack (IPv4 and IPv6)
endpoints by default. To preserve this behavior, enable this
value by default if it is not specified.

`DualStack` needs to be disabled if using an S3 Express bucket, since
this is not yet supported by AWS.

Previously S3 Transfer Acceleration could be enabled by setting
`ServerAddress`, but this can now be done with a single boolean.