Fix docker+machine download url
`development docker images` jobs haves started to fail recently (e.g.
https://gitlab.com/gitlab-org/gitlab-runner/-/jobs/7802187425). Turns
out the reason is that the way the release artifacts are hosted appears
to have changed.

The original download link

    "https://gitlab.com/gitlab-org/ci-cd/docker-machine/-/releases/${DOCKER_MACHINE_VERSION}/downloads/docker-machine-Linux-${DOCKER_MACHINE_ARCH}"

now return a redirect notice:

    <div class="tree-holder">
    <h2>You are being redirected away from GitLab</h2>
    <p>Redirect url is an external url, it may contain user-generated content and malicious code. Do not continue unless you trust the author and source.</p>
    </div>
    <div>
    <a href="%{redirect_uri}" target="_blank" rel="noopener noreferrer"> Click here to redirect to https://gitlab-docker-machine-downloads.s3.amazonaws.com/v0.16.2-gitlab.28/docker-machine-Linux-x86_64 </a>
    </div>

Fix the error by updating the download link to the link in the redirect notice.

I'm not convinced this is the right way to fix this (really the original
behaviour should be restored), but this will get our pipeline working
again in time for the upcoming release.

NOTE: This change should be reverted when the actual bug is fixed.