Docker executor integration tests
The goal here is to make sure we can execute steps in the various ways
they can be sent from GitLab:
- script step.
- local step.
- remote step.
- make sure file variables work.
- make sure job variables are not injected into the build container.

We'd also like a job to run github action, but I have not been able to
make that one work yet.

Note that in the near future, these test might be flaky since we don't
version the step-runner image, and changes there might brake tests here.
To deal with that we skip the test if the variable CI_SKIP_STEPS_TESTS
is set.