Refactor Azure cache implementation
This refactor defines a sasSigner interface that better reflects
how SAS token generation is performed with Azure and improves
the testing:

1. `accountKeySigner`
2. `userDelegationKeySigner`

`accountKeySigner` is the original version that uses static
credentials (account name and account key).

`userDelegationKeySigner` is the new version used for Azure workflow
and managed identities. A user delegation key is retrieved using the
Azure credentials.