Don't store creds in the git config
Instead of adding any creds to any git configuration, we get the token
from the environment, by configuring git to use a credential helper.
Furthermore, we only set this credential helper up in the helper
container (git config --global), so it is not available in the build
container.

However, the build container still has access to the token, it is set in
the environment of a step script.