Dialer implementation that tunnels though docker exec API
This code is absolutely heinous, and my grasp of it is fleeting. But it
is what is required to tunnel gRPC calls through a text-based docker
exec API. It satisfies the Dialer interface in
https://gitlab.com/gitlab-org/step-runner/-/blob/main/pkg/api/client/extended/client.go

It's mostly an exercise in satisfying interfaces and matching function
arguments. On the one end we need a

    func(context.Context, string) (net.Conn, error)

to pass to grpc.WithContextDialer() in grpc.ClientConn(). On the other
side we have a types.HijackedResponse returned from
ContainerExecAttach(). Most of this code is just reconciling those two
types.

We capture the container's stderr in an omitWriter to augment
stdcopy.StdCopy error messages.

This is a temporary home for this code; it will eventually be moved to a
new step-runner-client project.