Configure the git credential helper for the local repo
We don't configure the credential helper globally, but only for the
local repo. Side effect is, that not only the helper, but e.g. also the
build container has access to that. However, there is no sensitive data
in that config, merely a reference to the `$CI_JOB_TOKEN` environment
variable.