Method to determine if native steps integration is requested/enabled
A few convenience functions to be used elsewhere that will ultimately
tell us if we should be executing a job using native steps, and also
answer other finer grained questions to answer the larger question.

Native steps execution is enabled if:
- the job uses the run keyword.
- the feature flag is enabled.
- the executor supports native steps.

We'll use all of these later on when wiring up native steps execution.