Adapt git cred helper for all shells
Turns out, powershell does not like `-CommandWithArgs`, thus we are
moving away from that and just use `-Command`.

All cred helpers, regardless of the shell, now return with a 0 exit
code, when called with supported arguments ('get', '').

Also, a test was added which actually runs the git cred helper in an
appropriate shell and asserts on its actual output.