Make invalid service image name a build error
Originally this method returned a plain error when an invalid image name
was specified. This is interpreted higher up in the call stack as a
system error which results in:
* the job being retried twice more.
* the job finally failing with a system error.

This error is a build configuration error and, the job should fail as
such, without retries. Make it so be returning a BuildError.