Fix home directory detection
With 452ac0a2eb7bdf2895467b2cdab95d5f253e8d48 we changed how HOME
directory detection and "fixing" is working, but now it's causing
runner to fail in cases it didn't before.

The change was done because previously we've been using a deprecated
function from the `homedir` package. It was, however, done wrong.

This commit reverts code to the previous version.