Improve helpers/cli/FixHOME
FixHOME was previously relying on an external dependency
that recently deprecated some of its parts. That forced
us to make some changes as in 452ac0a2eb7bdf2895467b2cdab95d5f253e8d48
which, unfortunately, created a regression.

This commit was started with an intention to add some tests
to helpers/cli/fixHOME so that we will not repeat the same
regression in the past. However, during a short discussion we
had about the problem, we've decided that the code we're relying
on is so simple that we could just bring it to runner's codebase
directly, shaping the FixHOME behavior in the way we expected
it to work and dropping the dependency usage.