Set gitlab-advanced-sast job to run on code changes
This job was added some time ago to the `Jobs/SAST.latest.gitlab-ci.yml`
template we include. Other jobs from this template have - in our
pipeline - updated rules, so that they are executed only on code
changes.

`gitlab-advanced-sast` didn't yet got that override on our side, which
causes this job to be fired also for documentation or changelog changes.
It's not needed there and it just makes the pipelines execution to be
longer in these cases.

This commit aligns the configuration of the new SAST job with the rest
jobs we've been including so far.