Change how we pass in the arg for the git cred helper on windows
Previously, we ensured that in our helper script, we had a function
"get" set up. We then used the argument that git calls this helper with
to run the function; ie. if the helper has been called with "get", our
get function gets called, if the helper has been called with "foobar"
nothing would happen, but an error would be produced that there is no
cmdlet "foobar". This is generally fine, because git would ignore that
git cred helper and try the next (if there is one).

However, it turns out that
- there is a powershell built-in "erase", which is an alias for
  "Remove-Item"
- depending on other helpers (e.g. GCM) and their setup, git might call
  our helper with "erase"

Thus we changed how we consume the argument passed in by git. For the
well-known cases we set up functions (e.g. `gch_store`) which do "the
right thing" (which might be a no-op). And we'll make the main
entrypoint function `helper`, wich calls either one of those function or
fails with an error.