Externalize git cred helper
When git pulls submodules, it does not consider the cred helper in the
mother repo, but only in cred helpers from the global or system scope.

Thus we set up the cred helper in a separate config file, and include
that in the mother repo's git config. Also, when manipulating git
submodules, we include the same helper config explicitly on the git call
(`-c include.path=...`).

To ensure the cred helper matches for the mother repo and the
submodules' repos, the cred helper is now set to match for the whole
host (ie. https://gitlab.com) instead of the repo specifically. While
this might not strictly be necessary¹, this still makes more sense, as
we not only want the cred helper to apply to the mother repo, but all
repos from the GL server.

¹ https://github.com/git-ecosystem/git-credential-manager/blob/main/docs/configuration.md#credentialusehttppath