Method to validate steps request, and determine how to execute it
This method extends and will replace StepShim. In addition to the
request validation StepShim already did, this method detects if native
steps request was request and is supported by the executor, and sets an
appropriate ScriptStep. Otherwise it did whatever else StepShim did.

I've also ursurped the existing test to test this new method instead of
StepShim.