Implement CommandWithStdin for shell writer
This allows to run external commands with arguments, and passing in
data to the command via Stdin; abstracted for different shells.

A classic approach like `echo 'foo' | bar` does work across (ba)sh.

Powershell/Pwsh, however, presents issues with that. The way data is
handled "in" the pipeline makes it hard to pass on data as is;
powershell/pwsh seems to add BOM, mess with encoding, ...

Thus this is abstracted away in the shell writer and a special handling
for the powershell/pwsh case is implemented.