Make pwsh/powershell git cred helper align with the one for bash
The credential helpers now work similarly across different shells:
They only react when called with "get"; anything else will be a no-op,
regardless if this is a "well-known" command or an unknown one.

Note: The test TestPowershellGetGitCredHelperCommand has been removed,
as it would end up being a reimplementation of
GetGitCredHelperCommand(). We however have coverage in the
TestGitCredHelper.