Use S3 dual-stack when ServerAddress set to default AWS S3 address
Previously if `ServerAddress` were set to to `s3.amazonaws.com`, the
dual-stack S3 endpoints were disabled since the AWS SDK does not allow
you to set a custom endpoint and enable the dual-stack option
simultaneously.

We had one customer that appears to rely on the dual-stack endpoint,
possibly because the IPv4 endpoint is firewalled or proxied. To make
existing configurations work, avoid setting a custom endpoint if
`ServerAddress` is the default AWS S3 endpoint.

Changelog: fixed