Warn user if no fleeting plugins to install
This commit adds a warning message to inform the user if there are no
plugins to be installed.

Users do not run `gitlab-runner fleeting install` without intending to
install plugins, so it is important to provide this feedback to users so
they can resolve on their end (e.g. configuration issues).