Allow integration tests access to the CI_JOB_TOKEN
To test pulling of private repos we need a token. The integration tests
run via splitic and with an allow list which does _not_ pass on
CI_JOB_TOKEN.

To give tests access to a token, specifically the CI_JOB_TOKEN of the
current run, we expose it to all integration tests as
OUTER_CI_JOB_TOKEN. This way we don't "pollute" the tests' environment
and potentially hide future auth issues, but still give tests access to
the token.