Install git-lfs in ubi image from upstream RPM repo
For the ubi-base image only, install git-lfs from the project's official
release RPM repository. Why do we need to do this?
1. It turns out git-lfs is subject to a lot of CVEs
2. While the CVEs are fixed upstream in a timely manner, redhat is very
   slow about updating it's gitl-lfs packages.

This leaves us constantly exposed to a large number of active CVEs
against git-lfs.

In general I'd rather not address this situation this way, but because
of the above two points, and because git-lfs is easy to install (it has
no additional dependencies beyond git, which is already installed), it
is feasible to do.

NOTE to everyone: DO NOT try to use this as a precedent for dealing with
other CVEs in a similar way. This a single exception.

Note that we disable gpgcheck on the git-lfs repo because the rpm
packagesa re signed with SHA1 which is no longer supported in redhat
(https://www.redhat.com/en/blog/rhel-security-sha-1-package-signatures-distrusted-rhel-9)