shells: Implement the use of git-clone(1) again
In the past, the GIT_STRATEGY 'clone' used to use git-clone(1), but at
some point we needed clone a specific reference. For example for merged
result pipelines we'd clone refs/merge-requests/<iid>/merge. It is not
possible with git-clone(1) a ref, so the implementation was rewritten to
use git-init(1) + git-fetch(1).

Now we've been working to upstream changes to Git and soon it will
support argument `--ref`. This option allows use to use git-clone(1) to
clone and checkout a specific reference.

Prepare the code to use git-clone(1) with argument `--revisio`. The
changes are behind feature flag FF_USE_GIT_NATIVE_CLONE, and only should
be used if a the helper image contains a Git version that supports
`--revision`.

Issue: https://gitlab.com/gitlab-org/gitlab-runner/-/issues/37947