Omit canonical ports for S3 endpoints
If `ServerAddress` included the `:443` suffix, then presigned URLs for
S3 may fail with 403 Forbidden because the AWS SDK appears to use the
full endpoint to sign URLs. This was reported in the AWS SDK
JavaScript issue tracker
(https://github.com/aws/aws-sdk-js/issues/4062), and it appears to be
an issue with Go as well.

Fix this by dropping the canonical ports for HTTP and HTTPS if
present.

Relates to https://gitlab.com/gitlab-org/gitlab-runner/-/issues/38282

Changelog: fixed