Invoke step-runner from $PATH instead of /
Over in https://gitlab.com/gitlab-org/step-runner/-/merge_requests/138
we made an alias from `/step-runner` to `/usr/bin/step-runner` in the
`step-runner` image. As the MR mentions, this is because over in
https://gitlab.com/gitlab-org/gitlab-runner/-/merge_requests/5069 we
state that step-runner must be in $PATH for runner steps integration to
work. To make that true we implemented
https://gitlab.com/gitlab-org/step-runner/-/merge_requests/138. Now that
`step-runner` is in $PATH, we can update the invocation here.

But why do we want to do this? The _other_ way users can use
`step-runner` is to create their own build image that includes the
`step-runner` binary. In this case, users are more likely to put the
binary somewhere in $PATH than at `/`. And if they did do, the current
shim approach to using steps would not work.