Fix path-style requests with Upload ARN functionality
Previously only the `hostname_immutable` flag was set on the GoCloud
URL for AWS when the upload ARN were used. This is not sufficent.
As explained in https://github.com/google/go-cloud/issues/3472,
we also need the `use_path_style` query parameter.

This should not affect most buckets because path-style access only
works on buckets before September 2020
(https://aws.amazon.com/blogs/aws/amazon-s3-path-deprecation-plan-the-rest-of-the-story/).
However, in the oft chance we have customers still using path-style
access with S3 with the newly-introduced Upload ARN functionality,
then this change will allow that to work.

Changelog: fixed