Fix test
Turns out this test failed, but in a way that revealed a flaw in in the
test: no logger instance was being passed to the NewOSCmd options. This
caused a nil pointer exception when the test triggered the code that
logs in `assignProcessToJobObject()`. Because that code logs and
swallows errors, the test has to capture the logs and make sure they are
empty.