Attempt to fix stub ssh server on windows
In the same way that we fixed windows process handling. Note that the
process handling code is now identical between these two files. I
resisted the urge to consolidate the code because the ssh stub server is
test code, and i didn't want to intertwine prod and test code in this
way.