Support ImageLoad for prebuilt images
For a long time Runner has supported exported helper images
("prebuilt" images).

These are loaded using the API equivalent of 'docker import'. We use
them in 2 ways:

  1. They're "prebuilt" before tests, which is where the odd name
     probably comes from. Because they're built and stored on disk, they
     can be loaded by our tests without us having to push them to a
     registry just for the sake of executing tests.
  2. Some Runner distributions (DEB, RPM) bundle the prebuild images so
     that they can be loaded into target job environments that might not
     have internet access to our public registries.

Unfortunately, the exported image is devoid of metadata, meaning we have
to hard code some data to effectively restore iti. This makes it hard to
refactor.

The alternative to this, which is the fix included here, is to use
"docker load", which does load the metadata and is supported by Windows.