Fix sorting of OS release versions
getDistroReleasesToPackage() relied on the call to packgecloud to return
results ordered. Specifically it assumed the list versions for a given
OS were in the same order that the releases were made. In other words,
more or less chronological. The results are actually ordered by release
name (probably the `index_name` field).

For most OS these two are in fact the same, but SLES and openSUSE are stupid
exceptions. openSUSE for example has the following releases

{ "id": 71,   "index_name": "13.1" }
{ "id": 72,   "index_name": "13.2" }
{ "id": 96,   "index_name": "15.0" }
{ "id": 97,   "index_name": "15.1" }
{ "id": 119,  "index_name": "15.2" }
{ "id": 120,  "index_name": "15.3" }
{ "id": 138,  "index_name": "15.4" }
{ "id": 165,  "index_name": "15.5" }
{ "id": 174,  "index_name": "15.6" }
{ "id": 186,  "index_name": "16.0" }
{ "id": 73,   "index_name": "42.1" }
{ "id": 79,   "index_name": "42.2" }
{ "id": 83,   "index_name": "42.3" }

Where the 42.x series of releases are newer than the 13.x series, but
older than the 15.x series, as indicated by the id field. SLES is equaly
stupid:

{ "id": 74,   "index_name": "11.4" }
{ "id": 75,   "index_name": "12.0" }
{ "id": 76,   "index_name": "12.1" }
{ "id": 80,   "index_name": "12.2" }
{ "id": 91,   "index_name": "12.3" }
{ "id": 110,  "index_name": "12.4" }
{ "id": 111,  "index_name": "12.5" }
{ "id": 98,   "index_name": "15.0" }
{ "id": 112,  "index_name": "15.1" }
{ "id": 121,  "index_name": "15.2" }
{ "id": 122,  "index_name": "15.3" }
{ "id": 139,  "index_name": "15.4" }
{ "id": 175,  "index_name": "15.5" }
{ "id": 187,  "index_name": "15.6" }

This commit fixes the ordering by sorting versions by the id. This will
require updating the oldestRelease and skipReleases consts, which I'll
do next.