Fix TestStackDumping test freezing
The test was freezing intermittently because there was a race
condition:

1. The test called `watchForGoroutinesDump`, which set up a signal
   handler for `USR1` that will send `true` value to a non-blocking
   channel (`dumpedCh`) when this happens.

2. The test then issued the `USR1` signal.

3. The test listened on `dumpedCh`.

If the `USR1` signal were handled and sent a signal to `dumpedCh`
before the test was ready in step 3, the test could end up waiting
forever.

Fix this by making it possible to use a blocking channel to avoid
this raciness.

Relates to https://gitlab.com/gitlab-org/gitlab-runner/-/issues/38325