Update steps version to 0.2.0
Steps v0.2.0 was released recently, and made some backward-incompatible
breaking changes. Because of the way the
registry.gitlab.com/gitlab-org/step-runner:v0 image is released (the
same v0 tag is reused for all releases), the steps integration
integration tests started failing with:

    ERROR: Job failed (system failure): container exec on "6b6d4cefa3ed71e4f97254736db4c05ff815032ae00fb9585688099c6235ab70" finished with: executing step request: running run request for job "0": rpc error: code = Unknown desc = loading step: reading steps "{}\n---\nsteps:\n- name: echo\n  script: echo ${{ env.FLIN_FLAN_FLON }}\n": validating step: jsonschema: '' does not validate with https://gitlab.com/gitlab-org/step-runner/schema/v1/step#/$ref/oneOf/0/required: missing properties: 'name', 'step'
This is because we're using a 0.1.0 version of the library, but a 0.2.0
of the binary (included in the above image)

Elsewhere I have made these tests to be skipped by setting the
CI_SKIP_STEPS_TESTS variable, but to fix this we have to upgrade to
v0.2.0 of the steps library.

With the 0.2.0 release, the step
https://gitlab.com/gitlab-org/ci-cd/runner-tools/echo-step was also
changed, and since we use that step in the test, we have to also update
the version we use to `@v4` from `@v3`.

I'll remove the CI_SKIP_STEPS_TESTS variable when this MR is merged.