Limit UploadARN session duration to 1 hour
Trying to use a session duration of longer than hour previously
resulted in:

```
Error when assuming the role The requested DurationSeconds exceeds the
1 hour session limit for roles assumed by role chaining.
```

As described in
https://docs.aws.amazon.com/IAM/latest/UserGuide/id_roles_manage-assume.html:

Using the credentials for one role to assume a different role is
called role chaining. When you use role chaining, your new
credentials are limited to a maximum duration of one hour.

Relates to
https://gitlab.com/gitlab-org/gitlab-runner/-/issues/38370

Changelog: fixed