Implement a pod watcher
This new pod watcher uses an informer to get updates from the kube api.
It then does some checks on the pod & containers to figure out, if the
pod is still healthy / ready to run builds.

It uses a "private" informer factory for that, which it creates on
Start() and destroys on Stop().