Add new ConfigurationError failure reason
None of the existing failure reasons quite captures the case we want to
identify, which is incompatible pull policies. This is one case of the
more general error category of bad configuration, so I've added a new
failure reason to that end.

This new failure reason does not exist in rails (yet?) so I've mapped it
to ScriptFailure, which is the most correct of the exciting failures in
that it potentially indicates a user error.