PodWatcher: Gracefully handle initial cache sync
We run the initial cache sync of the pod watcher's informer with a
timeout, based on Config.Kubernetes.RequestRetryBackoffMax. If the
initial cache sync does not succeed in time, we bail out.

For issues at runtime, e.g. network disruption, we don't do anything and
let the client machinery do the back-off and retry.