CI job to ensure the docs are up to date
This is mostly a blatant copy/pasta of the k8s docs job.

The trigger here is a scheduled pipeline/job that runs against main.
This job will only:
- on main
- from a scheduled trigger
- in the canonical repo

The schedule is weekly on Monday mornings MST.

Note also the environment configuration; this is identical to the
packagecloud push and yank jobs.