Use a dotenv file to store cache environment variables
Previously when generating temporary credentials for Azure or AWS
(when RoleARN is set), the runner managed passed these credentials via
environment variables to the cache archiver/extractor.

This works, but a slightly better approach would be to load the
credentials only when the cache process needs to upload/download
files. This commit now:

1. Save them to a dotenv file.
2. When the process needs the credentials, load them into the
   environment.
3. Remove the dotenv.

This prevents the leaking of the tokens by variable expansion as
described in
https://gitlab.com/gitlab-org/gitlab-runner/-/issues/38475.

Changelog: changed