Add response detail debugging for artifact upload
The uploader uses the well known CI_DEBUG_TRACE to log more details
about the artifact upload response.
Those details are:
- response headers
- raw response body (truncated, if too long)

This is done, so that it's easier for users to debug when they hit
upload issues. E.g. proxies in front of the GitLab instance might add /
overwrite response headers or even the response body, which can be
discovered by the user now.