Introduce changes to custom exec build failure
This effort introduces optional behavior to the BUILD_FAILURE_EXIT_CODE
that aims to allow user of the custom executor to capture the real exit
code. When the BUILD_EXIT_CODE_FILE is created the GitLab runner will
retrieve the supplied exit code for the underlying build failure. The
primary goal of this is to allow jobs to leverage the
allow_failure:exit_code feature in conjunction with a custom executor.