Method to get step-runner image name:version
This method will tell us from which image:version to get the step-runner
binary to inject into the build environment. The default image is the
official `registry.gitlab.come/gitlab-org/step-runner`, and the default
version is the version of the step-runner library against which the
runner binary as compiled. Both can be specific via Variables.

Note that there is currently a limitation in debug.ReadBuildInfo() where
it returns an empty dependency list. This applies only to tests and only
when the test is not in the main module, and will be fixed in Go release
1.24
(https://github.com/golang/go/commit/d79e6bec6389dfeeec84a64f283055090615bad1)

Only use the major.minor versions for default step-runner image version.