Add step-runner volume
This volume will be shared among the helper and build containers; we'll
copy the step-runner binary into this volume, and it will be visible by
the build container.

I've made the mount point /opt/step-runner since it's unlikely that this
will conflict with anyting; this applies to all
containers that mount this volume.

The mount point on the step-runner container doesn't really matter, but
for the build container I was hoping to mount the binary at
`/usr/bin/step-runner` so I would not have to modify the container's
PATH. Unfortunately AFAICT it is not possible to mount a file (vs a dir)
using the docker APIs (it is possible at the command line). Plus, while
assuming `/usr/bin` is in PATH, there's no guarantee that it is.