Update k8s analyzer and docs generator
This introduces a new `kubeAPI` flag:
If we find an annotation to contain `=http`, we don't consider this a
config or feature flag, but a doc link flag. The part before the `=` is
used as a link name, the part after as an URL to docs about it.

When rendered into the permission table, as markdown, we render the link
as markdown to local anchors (ie. linking to the local document). This
is done, so that we adhere to the markdown liniting rules.

When we render into a code block, for the role yaml, we render the link
as "plain text".

This is all done, to make it easy for users to differentiate between
feature/config flags and "other important things", ie. those new doc
links.