Run the step-runner container
To keep things simple, I've decided to tie running the step-runner
container to the helper container since things currently done by the
helper container will eventually be done via steps.

Note that the helper container is "run" multiple times; once each to do
things like cloning the project, downloading the cache, downloading
artifacts.


A consequence of tying the step-runner container execution to the helper
container execution is that the step-runner container would be run
multiple times. While slightly wasteful, this would be OK because with
https://gitlab.com/gitlab-org/step-runner/-/merge_requests/170 the
bootstrap command is idempotent. This resolved in a cluttered job trace
though, so I wrapped the call in a sync.Once.