Document apt limitation and required workaround
apt/apt-get is broken: https://www.google.com/search?hl=en&q=apt%20dependency%20version%20bug

Specifically, when a package as a required dependency of a specific
version, and a version other than the latest is installed, apt will
contradict itself and ultimately refuse not install the package.

The workaround is to explicitly install the dependency at the required
version at the same time as the package.