Handle nil data at valid key for kv_v2 value engine
https://gitlab.com/gitlab-org/gitlab-runner/-/issues/38535 describes a
scenario where a vault key exists, but the secret has been marked
deleted and appears as `nil` data in the the kv_v2 engine. In this case
the engine panics because it can't coerce `nil` to a `map[string]any`.

Handle this case explicitly and return an error instead of panicing.