Add clear-docker-cache script to runner image
In the great image refactor of 18.0, we inadvertently dropped the clear-docker-cache script from the runner images.

Arguably the base images is a better place to put this, however, this
script must exist in this repo because it is also included in the runner
OS packages, and I do not want to duplicate it in another repo.

This addresses https://gitlab.com/gitlab-org/gitlab-runner/-/issues/38571